+++ packages/zowe-explorer/__mocks__/mockCreators/shared.ts
@@ -13,10 +13,11 @@ import { ZoweTreeProvider } from "../../src/abstract/ZoweTreeProvider";
 import { ZoweDatasetNode } from "../../src/dataset/ZoweDatasetNode";
 import { ZoweUSSNode } from "../../src/uss/ZoweUSSNode";
 import * as vscode from "vscode";
-import { ValidProfileEnum } from "@zowe/zowe-explorer-api";
+import { ValidProfileEnum, IZoweTreeNode } from "@zowe/zowe-explorer-api";
 import { FilterDescriptor } from "../../src/utils/ProfilesUtils";
 import { imperative, ZosmfSession } from "@zowe/cli";
 import { SettingsConfig } from "../../src/utils/SettingsConfig";
+import * as globals from "../../src/globals";
 
 export function createPersistentConfig() {
     return {
@@ -547,4 +548,44 @@ export function createOutputChannel() {
         dispose: jest.fn(),
         replace: jest.fn(),
     } as vscode.OutputChannel;
+}
+
+export function createMockNode(name: string, context: string): IZoweTreeNode {
+    return {
+        dirty: false,
+        getLabel: jest.fn(() => name),
+        getChildren: jest.fn(),
+        getParent: jest.fn(),
+        getProfile: jest.fn(),
+        getProfileName: jest.fn(),
+        getSession: jest.fn(),
+        getSessionNode: jest.fn(),
+        setProfileToChoice: jest.fn(),
+        setSessionToChoice: jest.fn(),
+        label: name,
+        contextValue: context,
+    };
+}
+
+export function createTreeProviders() {
+    return {
+        ds: {
+            mSessionNodes: [createMockNode("zosmf", globals.DS_SESSION_CONTEXT), createMockNode("zosmf2", globals.DS_SESSION_CONTEXT)],
+            deleteSession: jest.fn(),
+            removeSession: jest.fn(),
+            refresh: jest.fn(),
+        } as any,
+        uss: {
+            mSessionNodes: [createMockNode("zosmf", globals.USS_SESSION_CONTEXT), createMockNode("zosmf2", globals.USS_SESSION_CONTEXT)],
+            deleteSession: jest.fn(),
+            removeSession: jest.fn(),
+            refresh: jest.fn(),
+        } as any,
+        job: {
+            mSessionNodes: [createMockNode("zosmf", globals.JOBS_SESSION_CONTEXT), createMockNode("zosmf2", globals.JOBS_SESSION_CONTEXT)],
+            removeSession: jest.fn(),
+            deleteSession: jest.fn(),
+            refresh: jest.fn(),
+        } as any,
+    };
 }