+++ packages/zowe-explorer/src/shared/init.ts
@@ -29,6 +29,9 @@ import { SettingsConfig } from "../utils/SettingsConfig";
 import { spoolFilePollEvent } from "../job/actions";
 import { HistoryView } from "./HistoryView";
 import { ProfileManagement } from "../utils/ProfileManagement";
+import { DatasetTree } from "../dataset/DatasetTree";
+import { USSTree } from "../uss/USSTree";
+import { ZosJobsProvider } from "../job/ZosJobsProvider";
 
 // Set up localization
 nls.config({
@@ -38,10 +41,9 @@ nls.config({
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
 
 export interface IZoweProviders {
-    ds: IZoweTree<IZoweTreeNode>;
-    uss: IZoweTree<IZoweTreeNode>;
-    job: IZoweTree<IZoweTreeNode>;
-    // [key: string]: IZoweTree<IZoweTreeNode>;
+    ds: DatasetTree;
+    uss: USSTree;
+    job: ZosJobsProvider;
 }
 
 export function registerRefreshCommand(