+++ packages/zowe-explorer/src/uss/USSTree.ts
@@ -64,6 +64,7 @@ export class USSTree extends ZoweTreeProvider implements IZoweTree<IZoweUSSTreeN
     public mFavorites: IZoweUSSTreeNode[] = [];
     public lastOpened: NodeInteraction = {};
     private treeView: vscode.TreeView<IZoweUSSTreeNode>;
+    public copying: Promise<unknown>;
 
     public constructor() {
         super(
@@ -363,11 +364,9 @@ export class USSTree extends ZoweTreeProvider implements IZoweTree<IZoweUSSTreeN
      *
      * @param {IZoweUSSTreeNode} [node]
      */
-    public deleteSession(node: IZoweUSSTreeNode): void {
+    public deleteSession(node: IZoweUSSTreeNode, hideFromAllTrees?: boolean): void {
         ZoweLogger.trace("USSTree.deleteSession called.");
-        this.mSessionNodes = this.mSessionNodes.filter((tempNode) => tempNode.label.toString() !== node.label.toString());
-        this.mHistory.removeSession(node.label as string);
-        this.refresh();
+        super.deleteSession(node, hideFromAllTrees);
     }
 
     /**
@@ -854,6 +853,11 @@ export class USSTree extends ZoweTreeProvider implements IZoweTree<IZoweUSSTreeN
     public removeSearchHistory(name: string): void {
         ZoweLogger.trace("USSTree.removeSearchHistory called.");
         this.mHistory.removeSearchHistory(name);
+    }
+
+    public removeSession(name: string): void {
+        ZoweLogger.trace("USSTree.removeSession called.");
+        this.mHistory.removeSession(name);
     }
 
     public resetSearchHistory(): void {